fix(core): resolve data integrity, null safety, and logic errors
- settings-manager: parseInt -> parseFloat for decimal settings
- enhanced-data-manager: || vs ?? for isAuthenticated, retry re-add
  to pendingUserInfo, archive before delete in 4 cleanup paths
- api-client: null guards on 7 message handler paths, try/catch wraps
- content.js: store MutationObserver reference for cleanup
- heatmap-chart: guard Math.max on empty arrays
- creation-chart: division by zero guard
- stats-manager: Math.max(0, ...) for negative authenticatedNonBots
- format-utils: null/NaN guards on formatNumber and formatBytes
- viewer.js: NaN guard on formatDuration, null checks on dates/elements

+++ content/core/settings-manager.js
@@ -316,7 +316,7 @@ window.SettingsManager = class SettingsManager {
 
             switch (schema.type) {
                 case 'number':
-                    cleanValue = parseInt(value);
+                    cleanValue = parseFloat(value);
                     if (isNaN(cleanValue)) {
                         console.warn(`Invalid number for ${key}, using default`);
                         cleanValue = schema.default;
@@ -365,7 +365,7 @@ window.SettingsManager = class SettingsManager {
 
         // Type check
         if (schema.type === 'number') {
-            const num = parseInt(value);
+            const num = parseFloat(value);
             if (isNaN(num)) return { valid: false, error: 'Must be a number' };
             if (schema.min !== undefined && num < schema.min) {
                 return { valid: false, error: `Must be at least ${schema.min}` };
